Login Query
Made big changes to the login query. Included the user image path from
the system_users table from the database. Also when user login, s/he
will see the profile picture at the top right corner of the home page.
The path of the profile picture is stored in a session. Also disabled
the default right click to trigger the footer to toggle up and down.