My Task Working
Successfully implemented the interface to allow the user to view the
tasks that s/he created and the user can also make a preview of those
tasks to see more details about the task. Also implemented the search
for 'My Task'. Thus the user can search for a task that s/he created.
About to make an implementation that will allow the user to delete a
task that he created.